% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furcation-class.R
\docType{class}
\name{allelefurcation-class}
\alias{allelefurcation-class}
\alias{allelefurcation}
\title{An S4 class containing furcation trees for one allele
of a focal marker}
\description{
An S4 class containing the furcation trees
for both sides of a focal marker for one allele.
}
\section{Slots}{

\describe{
\item{\code{allele}}{the allele of the focal marker.}

\item{\code{description}}{"ancestral", "derived", "major", "minor", etc.}

\item{\code{count}}{the number of chromosomes with that allele.}

\item{\code{left}}{furcation tree to the left of the marker.}

\item{\code{right}}{furcation tree to the right of the marker.}
}}

\seealso{
\code{\link{ftree}}, \code{\link{furcation}}
}
