\name{hap_cgu_bta12}
\Rdversion{1.1}
\alias{hap_cgu_bta12}
\docType{data}

\title{Haplotype data from a New World Creole cattle breed}


\description{
  A matrix of 280 rows (haplotypes) and 1424 columns (SNPs mapping to chromosome BTA12).
}

\usage{data(hap_cgu_bta12)}

\references{
   Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World
   Creole cattle breed. \emph{Molecular Ecology}, \bold{20}, 3128--3143.
}

