% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplohh-class.R
\docType{class}
\name{haplohh-class}
\alias{haplohh-class}
\alias{chr.name}
\alias{positions}
\alias{haplo}
\alias{nmrk}
\alias{nhap}
\alias{mrk.names}
\alias{hap.names}
\alias{chr.name,haplohh-method}
\alias{chr.name,}
\alias{positions,haplohh-method}
\alias{positions,}
\alias{haplo,haplohh-method}
\alias{haplo,}
\alias{nmrk,haplohh-method}
\alias{nmrk,}
\alias{mrk.names,haplohh-method}
\alias{mrk.names,}
\alias{nhap,haplohh-method}
\alias{nhap,}
\alias{hap.names,haplohh-method}
\alias{hap.names,}
\title{Class "haplohh"}
\usage{
\S4method{chr.name}{haplohh}(x)

\S4method{positions}{haplohh}(x)

\S4method{haplo}{haplohh}(x)

\S4method{nmrk}{haplohh}(x)

\S4method{mrk.names}{haplohh}(x)

\S4method{nhap}{haplohh}(x)

\S4method{hap.names}{haplohh}(x)
}
\arguments{
\item{x}{an object of this class.}
}
\description{
An object of this class contains the information needed for computation of EHH based statistics.
}
\details{
This class is the basis for all calculations done by this package.
Note that the matrix in slot \code{haplo} has to be of type \code{integer}, not \code{numeric}.
Objects built by versions of rehh up to 2.0.4 coded this matrix as \code{numeric} and used
a different coding scheme. They can be converted e.g. by
\code{haplohh <- update_haplohh(old_haplohh)} in order be used
with the present version.
}
\section{Slots}{

\describe{
\item{\code{chr.name}}{name of the chromosome/scaffold to which the markers belong.}

\item{\code{positions}}{vector of type \code{numeric} containing the marker positions within the chromosome.}

\item{\code{haplo}}{matrix of type \code{integer} containing haplotypes in rows and markers in columns.}
}}

\examples{
showClass("haplohh")
}
\seealso{
\code{\link{data2haplohh}}, \code{\link{update_haplohh}}
}
