% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.example.files.R
\docType{data}
\name{haplohh_cgu_bta12}
\alias{haplohh_cgu_bta12}
\title{Example of an \code{haplohh} object}
\format{
An object of \code{\link{haplohh-class}}.
}
\usage{
data(haplohh_cgu_bta12)
}
\description{
The object contains haplotype data for 140 cattle individuals (280 haplotypes) belonging to the
Creole breed from Guadeloupe (CGU) and 1424 markers (mapping to chromosome BTA12).
}
\references{
Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World Creole cattle breed. \emph{Molecular Ecology}, \strong{20}, 3128-3143.
}
\seealso{
\code{\link{data2haplohh}}
}
\keyword{datasets}
