% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanplot.R
\name{manhattanplot}
\alias{manhattanplot}
\title{Manhattan plot of iHS, XP-EHH or Rsb over a genome.}
\usage{
manhattanplot(data, pval = FALSE, threshold = c(-2, 2),
  chr.name = NA, cr = NULL, cr.col = "gray", cr.opacity = 0.5,
  cr.lab.cex = 0.6, cr.lab.offset = 0, cr.lab.pos = "top",
  main = NA, xlim = NULL, cex = 0.5, las = 1, pch = 20, ...)
}
\arguments{
\item{data}{output of either \code{\link{ihh2ihs}}, \code{\link{ies2xpehh}} or \code{\link{ines2rsb}}.}

\item{pval}{logical. If \code{TRUE}, the p-value is plotted, otherwise the score itself.}

\item{threshold}{a horizontal line is added at the corresponding value(s),
for instance to represent a significance threshold.
A single value (upper or lower threshold) or two values (upper and lower) can be specified.}

\item{chr.name}{if \code{NA} (default), all chromosomes are plotted, otherwise only those specified.}

\item{cr}{highlight "candidate regions" specified by a data.frame with three
columns: the first containing the chromosome, the other begin and end of the region as
obtained by the function \code{\link{calc_candidate_regions}}.}

\item{cr.col}{the color for highlighting}

\item{cr.opacity}{a value between 0 (invisible) and 1 (opaque).}

\item{cr.lab.cex}{text size of candidate region labels.}

\item{cr.lab.offset}{offset of candidate region labels.}

\item{cr.lab.pos}{if \code{"top"} (default) or \code{"bottom"}, candidate regions are labeled by numbers; to turn off, use \code{"none"}}

\item{main}{main title of the plot.}

\item{xlim}{set x coordinate range of the plot. Ignored, if more than one chromosome is depicted.}

\item{cex}{size of the points representing markers in the plot(s) (see \code{\link[graphics]{par}}).}

\item{las}{orientation of axis labels (see \code{\link[graphics]{par}}).}

\item{pch}{type of the points representing markers in the plot(s) (see \code{\link[graphics]{points}}).}

\item{...}{further arguments to be passed to \code{\link[graphics]{plot.default}}.}
}
\value{
The function returns a plot.
}
\description{
Manhattanplot of iHS, XP-EHH or Rsb over a genome.
}
\details{
The color of chromosomes is taken from the "Graphics Palette", see \code{\link[grDevices]{palette}}.
}
\examples{
library(rehh.data)
data(wgscan.cgu)
## results from a genome scan (44,057 SNPs)
## see ?wgscan.eut and ?wgscan.cgu for details
wgscan.ihs <- ihh2ihs(wgscan.cgu)
manhattanplot(wgscan.ihs)
}
\seealso{
\code{\link{ihh2ihs}}, \code{\link{ies2xpehh}}, \code{\link{ines2rsb}}, \code{\link{calc_candidate_regions}}.
}
