\name{map_bta12}
\Rdversion{1.1}
\alias{map_bta12}
\docType{data}

\title{Map for SNPs on cattle chromosome BTA12}


\description{
  A matrix of 1424 rows (SNPs) and 2 columns (Chromosome name and position of the SNP in pb).
}

\usage{data(map_bta12)}

\references{
   Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World
   Creole cattle breed. \emph{Molecular Ecology}, \bold{20}, 3128--3143.
}

