% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.example.files.R
\name{remove.example.files}
\alias{remove.example.files}
\title{Remove example files from current working directory.}
\usage{
remove.example.files()
}
\description{
Remove example files from current working directory.
}
\details{
Removes the files created by \code{make.example.files()}. 
No error is thrown, if files do not exist.
}
\seealso{
\code{\link{make.example.files}}
}
