% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_haplohh.R
\name{update_haplohh}
\alias{update_haplohh}
\title{Update object of class haplohh}
\usage{
update_haplohh(haplohh)
}
\arguments{
\item{haplohh}{an object of an old version of \code{\link{haplohh-class}}.}
}
\description{
Update object of class \code{\link{haplohh-class}}
constructed by rehh versions up to version 2.0.4.
}
\details{
This function is intended to update \code{haplohh} objects
that have been built by rehh versions up to 2.0.4. These objects cannot
be used in functions of the current version. 
The following changes have been made to the class definition:
The internal representation of the haplotype matrix followed the encoding
\itemize{
\item 0 missing value
\item 1 ancestral allele
\item 2 derived allele}
and has been replaced by a vcf-like encoding:
\itemize{
\item \code{NA} missing value
\item 0 ancestral allele
\item 1 derived allele.}
Furthermore the slots \code{nsnp}, \code{snp.name} and \code{nhap} have been removed
and slot \code{position} renamed to \code{positions}.
An update of an old \code{haplohh} object is done as follows:

\code{new_haplohh = update_haplohh(old_haplohh)}.
}
\seealso{
\code{\link{haplohh-class}}, \code{\link{data2haplohh}}.
}
