\name{wgscan.eut}
\alias{wgscan.eut}
\docType{data}

\title{Whole genome scan results for a pool of European taurine cattle}


\description{
  A dataframe object with 44,057 rows (SNPs) and 7 columns: i) chromosome name (CHR), ii) position of the SNP in bp (POSITION), iii) Ancestral allele frequency (freq_A), iv) iHH for the ancestral allele (iHH_A), v) iHH for the derived allele (iHH_D), vi) iES using the estimator by Tang et al. (2007) (iES_Tang_et_al_2007), vii) iES using the estimator by Sabeti et al. (2007) (iES_Sabeti_et_al_2007).
}

\usage{data(wgscan.eut)}

\references{
   Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World
   Creole cattle breed. \emph{Molecular Ecology}, \bold{20}, 3128--3143.
}

