\name{haplohh_cgu_bta12}
\alias{haplohh_cgu_bta12}
\docType{data}

\title{Example of an \code{haplohh} object}


\description{
  The object contains haplotype data for 140 cattle individuals (280 haplotypes) belonging to the Creole breed from Guadeloupe (CGU) and 1424 SNPs (mapping to chromosome BTA12).
}
\seealso{data2haplohh}
\usage{data(haplohh_cgu_bta12)}

\references{
   Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World
   Creole cattle breed. \emph{Molecular Ecology}, \bold{20}, 3128--3143.
}

