\name{wgscan.eut}
\alias{wgscan.eut}
\docType{data}

\title{Whole genome scan results for a pooled of European taurine cattle}


\description{
  A matrix of 44,057 rows (SNPs) and 6 columns (Chromosome name, position of the SNP in bp, Frequency of the ancestral allele, iHH for the ancestral allele, iHH for the derived allele and iES).
}

\usage{data(wgscan.eut)}

\references{
   Gautier, M. and Naves, M. (2011). Footprints of selection in the ancestral admixture of a New World
   Creole cattle breed. \emph{Molecular Ecology}, \bold{20}, 3128--3143.
}

